requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
pystray>=0.19.0
Pillow>=10.0.0
